+++ Validator/NumberMaximumValidator.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Linkin\Bundle\SwaggerResolverBundle\Validator;
 
 use EXSyst\Component\Swagger\Schema;
+use Linkin\Bundle\SwaggerResolverBundle\Enum\ParameterTypeEnum;
 use Symfony\Component\OptionsResolver\Exception\InvalidOptionsException;
 use function in_array;
 use function sprintf;
@@ -28,7 +29,9 @@ class NumberMaximumValidator implements SwaggerValidatorInterface
      */
     public function supports(Schema $property, array $context = []): bool
     {
-        return in_array($property->getType(), ['number', 'integer'], true) && null !== $property->getMaximum();
+        return in_array($property->getType(), [ParameterTypeEnum::NUMBER, ParameterTypeEnum::INTEGER], true)
+            && null !== $property->getMaximum()
+        ;
     }
 
     /**