+++ Validator/StringMinLengthValidator.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Linkin\Bundle\SwaggerResolverBundle\Validator;
 
 use EXSyst\Component\Swagger\Schema;
+use Linkin\Bundle\SwaggerResolverBundle\Enum\ParameterTypeEnum;
 use Symfony\Component\OptionsResolver\Exception\InvalidOptionsException;
 use function mb_strlen;
 use function sprintf;
@@ -28,7 +29,7 @@ class StringMinLengthValidator implements SwaggerValidatorInterface
      */
     public function supports(Schema $property, array $context = []): bool
     {
-        return 'string' === $property->getType() && null !== $property->getMinLength();
+        return ParameterTypeEnum::STRING === $property->getType() && null !== $property->getMinLength();
     }
 
     /**